Fix install.sh

+++ bash/install.sh
@@ -1,13 +1,14 @@
 #!/usr/bin/env bash
 
+# Run with source install.sh in an admin terminal
 # Get the directory where the script is located
 BASEDIR="$(cd "$(dirname -- "$0")" && pwd)"
 cd "$BASEDIR" || exit 1  # Exit if changing directory fails
 
 # Remove existing configuration files if they exist
-rm ~/.bashrc
-rm ~/.vimrc
-rm ~/.pylintrc
+[ -f ~/.bashrc ] && rm ~/.bashrc
+[ -f ~/.vimrc ] && rm ~/.vimrc
+[ -f ~/.pylintrc ] && rm ~/.pylintrc
 
 # Create symbolic links to the new configuration files
 # Check if the link already exists before creating it
